refactor: 절대경로 설정 및 import 정리, 디렉토리 정리
