+++ src/App.jsx
@@ -1,9 +1,9 @@
 import { useState } from "react";
-import GlobalStyle from "./globalStyles.jsx";
+import GlobalStyle from "globalStyles.jsx";
 
-import CardAddPage from "./pages/CardAddPage.jsx";
-import CardListPage from "./pages/CardListPage.jsx";
-import { PAGE_NAME } from "./utils/constants.js";
+import CardAddPage from "pages/CardAddPage.jsx";
+import CardListPage from "pages/CardListPage.jsx";
+import { PAGE_NAME } from "utils/constants.js";
 
 function App() {
   const [page, setPage] = useState(PAGE_NAME.CARD_LIST);
