+++ src/components/CardNumberInput.jsx
@@ -1,8 +1,10 @@
 import React, { useCallback, useContext, useState } from "react";
-import { CARD_INFO_RULES, CREATE_MASKED_CHARACTERS } from "../utils/constants";
-import CardInfoContext from "../context/CardInfoContext";
-import Input from "./UIComponents/Input/Input";
-import InputField from "./UIComponents/InputField/InputField";
+import CardInfoContext from "context/CardInfoContext";
+
+import { CARD_INFO_RULES, CREATE_MASKED_CHARACTERS } from "utils/constants";
+
+import Input from "components/UIComponents/Input/Input";
+import InputField from "components/UIComponents/InputField/InputField";
 
 export default function CardNumberInput() {
   const [isInvalid, setInvalid] = useState(false);
