+++ src/components/UIComponents/Form/Form.jsx
@@ -1,6 +1,7 @@
 import React, { useEffect, useRef, useState } from "react";
 import styled from "styled-components";
-import { cardInfoValidations } from "../../../utils/cardInfoValidations.js";
+
+import { cardInfoValidations } from "utils/cardInfoValidations.js";
 
 const StyledCardInfoForm = styled.form`
   display: flex;
