+++ src/components/UIComponents/WordCounter/WordCounter.jsx
@@ -1,6 +1,7 @@
 import React from "react";
 import styled from "styled-components";
-import { textColor } from "../styleConstants";
+
+import { textColor } from "components/UIComponents/styleConstants";
 
 const StyledWordCounter = styled.p`
   position: absolute;
